Move test to correct position and updated tests
